Move abstract model definitions into an abstract_models module
This allows us to create the concrete models inside some other app, without modelsearch having to be included in INSTALLED_APPS.
